@@ -1,13 +1,13 @@
 {
   "name": "signalwerk.documentation.md",
-  "version": "1.0.0",
-  "description": "Content for publisher",
+  "type": "module",
+  "version": "2.0.0",
+  "description": "markdown processor",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server",
-    "lint": "eslint --ext .js --ignore-path .gitignore .",
-    "build": "NODE_ENV=production webpack && npm run gh-pages",
-    "gh-pages": "cp ./public/CNAME ./dist",
+    "start": "webpack-dev-server --config ./webpack.config.js",
+    "test-parse": "node ./src/md-parser/test/test.js",
+    "test-write": "node ./src/html-writer/test/test.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -19,40 +19,50 @@
     "md",
     "markdown"
   ],
-  "author": "Stefan Huber – sh@signalwerk.ch",
+  "author": "Stefan Huber",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/signalwerk/signalwerk.documentation.md/issues"
   },
   "homepage": "https://github.com/signalwerk/signalwerk.documentation.md#readme",
-  "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "babel-loader": "^8.0.6",
-    "copy-webpack-plugin": "^5.1.2",
-    "css-loader": "^3.2.0",
+  "dependencies": {
+    "escape-html-template-tag": "^2.1.0",
+    "front-matter": "^4.0.2",
+    "mdast-util-gfm-autolink-literal": "^1.0.2",
+    "mdast-util-gfm-strikethrough": "^1.0.0",
+    "mdast-util-gfm-table": "^1.0.2",
+    "mdast-util-gfm-task-list-item": "^1.0.0",
+    "micromark-extension-gfm-autolink-literal": "^1.0.3",
+    "micromark-extension-gfm-strikethrough": "^1.0.4",
+    "micromark-extension-gfm-table": "^1.0.5",
+    "micromark-extension-gfm-task-list-item": "^1.0.3",
+    "micromark-util-combine-extensions": "^1.0.0",
+    "prettier": "^2.5.1",
+    "rehype-external-links": "^1.0.1",
+    "rehype-highlight": "^5.0.2",
+    "rehype-raw": "^6.1.0",
+    "rehype-sanitize": "^5.0.1",
+    "rehype-stringify": "^9.0.2",
+    "remark-parse": "^10.0.1",
+    "remark-rehype": "^10.1.0",
+    "unified": "^10.1.1",
+    "unist-util-visit": "^4.1.0",
     "css-mqpacker": "^7.0.0",
-    "dateformat": "^4.6.3",
-    "eslint": "^6.2.2",
-    "eslint-config-prettier": "^6.1.0",
-    "eslint-plugin-prettier": "^3.1.0",
-    "front-matter": "^3.0.2",
-    "handlebars": "^4.7.7",
-    "highlight.js": "^10.4.1",
-    "marked": "^0.7.0",
-    "mini-css-extract-plugin": "^0.8.0",
-    "postcss-calc": "^7.0.1",
-    "postcss-conditionals": "^2.1.0",
+    "mini-css-extract-plugin": "^2.4.5",
+    "postcss": "^8.4.5",
+    "postcss-mixins": "^9.0.1",
+    "postcss-calc": "^8.0.0",
     "postcss-critical-split": "^2.5.3",
-    "postcss-import": "^12.0.1",
-    "postcss-loader": "^3.0.0",
-    "postcss-nested": "^4.1.2",
-    "postcss-pipeline-webpack-plugin": "^5.1.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-simple-vars": "^5.0.2",
-    "url-loader": "^2.1.0",
-    "webpack": "^4.39.2",
-    "webpack-cli": "^3.3.7",
-    "webpack-dev-server": "^3.8.0"
+    "postcss-import": "^14.0.2",
+    "postcss-loader": "^6.2.1",
+    "postcss-nested": "^5.0.6",
+    "postcss-pipeline-webpack-plugin": "^6.0.0",
+    "postcss-preset-env": "^7.1.0",
+    "postcss-simple-vars": "^6.0.3",
+    "showdown": "^1.9.1",
+    "css-loader": "^6.5.1",
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.7.1"
   }
 }
