@@ -1,82 +1,64 @@
-const path = require("path");
-const fs = require("fs");
-const webpack = require("webpack");
-const markdownPlugin = require("./webpack/md-to-html-plugin");
-const CopyPlugin = require("copy-webpack-plugin");
-const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+import { fileURLToPath } from "url";
+import { dirname, resolve } from "path";
+/*-------------------------------------------------*/
 
-// postcss plugins
-const postcss = require("postcss");
-const postcssImport = require("postcss-import");
-const postcssVars = require("postcss-simple-vars");
-const postcssCalc = require("postcss-calc");
-const postcssNested = require("postcss-nested");
-const postConditionals = require("postcss-conditionals");
-const postcssPresetEnv = require("postcss-preset-env");
-const PostCssPipelineWebpackPlugin = require("postcss-pipeline-webpack-plugin");
-const criticalSplit = require("postcss-critical-split");
-const postcssMqPacker = require("css-mqpacker");
+import postcss from "postcss";
+import postcssImport from "postcss-import";
+import postcssCalc from "postcss-calc";
+import postcssNested from "postcss-nested";
+import postcssPresetEnv from "postcss-preset-env";
+import criticalSplit from "postcss-critical-split";
+import postcssMqPacker from "css-mqpacker"; // Pack same CSS media query rules into one using PostCSS
+import postcssSimpleVars from "postcss-simple-vars";
+import postcssMixins from "postcss-mixins";
+import MiniCssExtractPlugin from "mini-css-extract-plugin";
+import PostCssPipelineWebpackPlugin from "postcss-pipeline-webpack-plugin";
+// import postConditionals from "postcss-conditionals";
+
+/*-------------------------------------------------*/
+
+import mdProcessor from "./src/webpack/md-processor.js";
+
+/*-------------------------------------------------*/
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+/*-------------------------------------------------*/
 
 const isProduction = process.env.NODE_ENV === "production";
 
-const root = "./"; // __dirname
-
-let localCSSsearch = "./src/main.css";
-let localJSsearch = "./src/index.js";
-let packageCSS = "./packages/signalwerk.documentation.md/root.css";
-let rootCSS = packageCSS;
-
-const aliasSearch = path.resolve(__dirname, "src/");
-
-if (fs.existsSync(path.resolve(root, localCSSsearch))) {
-  rootCSS = localCSSsearch;
-}
-
-const mainEntry = [path.resolve(root, rootCSS)];
-
-if (fs.existsSync(path.resolve(root, localJSsearch))) {
-  mainEntry.push(path.resolve(root, localJSsearch));
-}
-
-module.exports = {
+const settings = {
   // webpack optimization mode
   mode: process.env.NODE_ENV ? process.env.NODE_ENV : "development",
 
-  resolve: {
-    alias: {
-      "@": aliasSearch,
-    },
-    // extensions: [".css"],
-  },
-
   // entry file(s)
   entry: {
-    main: mainEntry,
+    root: ["./root.css"],
   },
 
   // output file(s) and chunks
   output: {
-    path: path.resolve(root, "dist"),
-    filename: "index.js",
+    path: resolve(__dirname, "dist"),
+    // filename: "index.js",
     publicPath: "/",
   },
 
   module: {
     rules: [
       {
-        test: /\.js$/,
-        exclude: /node_modules/,
-        use: {
-          loader: "babel-loader",
-        },
-      },
-      {
         test: /\.(jpg|jpeg|png|woff|otf|woff2|eot|ttf|svg)$/,
-        loader: "url-loader?limit=100000",
+        use: [
+          {
+            loader: "url-loader",
+            options: {
+              limit: 8192,
+            },
+          },
+        ],
       },
       {
         test: /\.css$/,
-
         use: [
           {
             loader: MiniCssExtractPlugin.loader,
@@ -87,29 +69,35 @@ module.exports = {
               sourceMap: isProduction ? false : true,
             },
           },
+
           // Add PostCSS
           {
             loader: "postcss-loader",
             options: {
-              ident: "postcss",
-              plugins() {
-                return [
-                  postcssImport(),
-                  postcssVars({
-                    variables: () => ({
-                      debug: isProduction ? 0 : 1,
-                    }),
+              postcssOptions: {
+                plugins: [
+                  postcssImport({
+                    // root: resolve(__dirname, "./Resources/Private"),
                   }),
+                  postcssMixins(),
+                  postcssSimpleVars({
+                    onVariables(variables) {
+                      // console.log("CSS Variables");
+                      // console.log(JSON.stringify(variables, null, 2));
+                    },
+                  }),
+
                   postcssCalc(),
                   postcssPresetEnv({
                     stage: 0,
                     browsers: ["last 2 versions", "IE > 10"],
                   }),
-                  postConditionals(),
+                  // postConditionals(),
                   postcssNested(),
                   postcssMqPacker(),
-                ];
+                ],
               },
+
               // Enable sourcemaps in development
               sourceMap: isProduction ? false : true,
             },
@@ -125,26 +113,12 @@ module.exports = {
     // new CopyPlugin({
     //   patterns: [
     //     {
-    //       from: path.resolve(root, "static"),
+    //       from: resolve(__dirname, "static"),
     //       to: "", // copies all files to dist
     //     },
     //   ],
     // }),
     // ---------------------------------------------------------
-
-    new CopyPlugin([
-      {
-        from: "static",
-        to: "", // copies all files to dist
-      },
-    ]),
-
-    new MiniCssExtractPlugin({
-      // Options similar to the same options in webpackOptions.output
-      // both options are optional
-      filename: "./styles/[name].css",
-      chunkFilename: "./styles/[id].css",
-    }),
 
     new PostCssPipelineWebpackPlugin({
       suffix: "critical",
@@ -154,6 +128,7 @@ module.exports = {
         }),
       ]),
     }),
+
     new PostCssPipelineWebpackPlugin({
       suffix: "rest",
       processor: postcss([
@@ -162,7 +137,13 @@ module.exports = {
         }),
       ]),
     }),
-    new markdownPlugin({
+
+    new MiniCssExtractPlugin({
+      filename: "./styles/[name].css",
+      chunkFilename: "./styles/[id].css",
+    }),
+
+    new mdProcessor({
       // filePath: "./content",
       // exportPath: "./dist/",
       // isEncodeName: false, // if need to encode file name, like chinese
@@ -172,12 +153,11 @@ module.exports = {
 
   // development server configuration
   devServer: {
-    contentBase: path.join(root, "dist"),
-    compress: true,
     port: 9001,
-    watchContentBase: true,
-
-    // open browser on server start
-    // open: true
+    allowedHosts: "all",
+    static: ["dist"],
+    // open: true,
   },
 };
+
+export default settings;
